Revert event changes
